Lots of things - docced in collab

+++ Assets/Scripts/Debug/BuildLog.cs
@@ -6,13 +6,17 @@ using UnityEngine;
 public class BuildLog
 {
 
+    public static bool isBuild = false;
+
     public static void writeLog(string log)
     {
+        if (!isBuild) return;
         File.AppendAllText("buildLogFile.txt", log + "\t\n");
     }
 
     public static void flushLogFile()
     {
+        if (!isBuild) return;
         File.Delete("buildLogFile.txt");
         File.WriteAllText("buildLogFile.txt", "");
     }
